% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vervet_data.R
\docType{data}
\name{df.ranks}
\alias{df.ranks}
\title{Ververt ranking estimates}
\format{An object of class \code{data.frame} with 62 rows and 3 columns.}
\source{
\href{https://www.biorxiv.org/content/10.1101/692384v1?rss=1}{QTL Archive}
}
\usage{
data(rank_df)
}
\description{
Ranking estimates from a Vervet population
}
\examples{
data(df.ranks)
df.ranks

}
\references{
Vilette et al. (2020) 
(\href{https://www.biorxiv.org/content/10.1101/692384v1?rss=1}{biorxiv})
}
\keyword{datasets}
