% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vervet_data.R
\docType{data}
\name{df.testing}
\alias{df.testing}
\title{Ververt dominance data}
\format{An object of class \code{data.frame} with 2980 rows and 3 columns.}
\source{
\href{https://www.biorxiv.org/content/10.1101/692384v1?rss=1}{QTL Archive}
}
\usage{
df.testing
}
\description{
Data from a Vervet population
}
\references{
Vilette et al. (2020) 
(\href{https://www.biorxiv.org/content/10.1101/692384v1?rss=1}{biorxiv})
}
\keyword{datasets}
