% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability_check}
\alias{reliability_check}
\title{Rank reliability estimate}
\usage{
reliability_check(rank.df, testing.df)
}
\arguments{
\item{rank.df}{A dataframe containing the identity in the first coloumn, with their rank (or rank score) in the second column. Note: additional rank estimates can be added to subsequent columns, e.g., id, elo_ranking, david_score_ranks,  ... etc.}

\item{testing.df}{A dataframe of dominance interactions: winner, loser, datetime}
}
\description{
Estimates the  reliability of rank using dominance data
}
