% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{reliability_plot}
\alias{reliability_plot}
\title{A plotting function for output from reliability_check()}
\usage{
reliability_plot(global.reliability, df.reliability, method = "loess")
}
\arguments{
\item{global.reliability}{A dataframe containing the estimates of overall reliability from the reliability_check() function.}

\item{df.reliability}{A dataframe containing the dyad level estimates of reliability from the reliability_check() function.}
}
\description{
This function creates a plot using the outputs from reliability_check()
}
