% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{set_testing}
\alias{set_testing}
\title{Create a testing dataset}
\usage{
set_testing(testing.raw, ties = "remove")
}
\arguments{
\item{testing.raw}{A dataframe containing the observed dominance data with the first and second coloumns containing the individual, a third column with the outcomes, and a fourth with a date (as ymd).}

\item{ties}{Wether to remove or keep ties/draws in the testing dataset.}
}
\description{
This function creates a testing dataset that can then be used by the function 'reliability_check'.
}
